requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.51.0
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-cors-headers==3.12.0
django-imagekit==4.1.0
djangorestframework==3.13.1
idna==3.3
oauthlib==3.2.0
pilkit==2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9